refactor: 결제 승인 기능 인터페이스 분리 후 에러 핸들러 추가
