+++ src/main/java/roomescape/exception/custom/reason/payment/PaymentException.java
@@ -1,0 +1,12 @@
+package roomescape.exception.custom.reason.payment;
+
+public class PaymentException extends RuntimeException {
+
+    public PaymentException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+
+    public PaymentException(final String message) {
+        super(message);
+    }
+}
