+++ src/main/java/roomescape/payment/PaymentClient.java
@@ -1,0 +1,8 @@
+package roomescape.payment;
+
+import roomescape.payment.dto.PaymentConfirmRequest;
+
+public interface PaymentClient {
+
+    void confirm(PaymentConfirmRequest request);
+}
