+++ src/main/java/roomescape/payment/dto/TossErrorResponse.java
@@ -1,0 +1,8 @@
+package roomescape.payment.dto;
+
+public record TossErrorResponse(
+        String code,
+        String message,
+        Object data
+) {
+}
